Apply Mareks patch
