+++ chatkit-core/src/test/kotlin/com/pusher/chatkit/rooms/GetJoinableRoomsParserSpec.kt
@@ -7,24 +7,23 @@ import com.pusher.chatkit.util.parseAs
 import org.spekframework.spek2.Spek
 import org.spekframework.spek2.style.specification.describe
 
-object FetchRoomsParserSpec : Spek({
+object GetJoinableRoomsParserSpec : Spek({
     val testFileReader = TestFileReader("/json/rooms")
 
-    describe("when fetching joinable rooms") {
-
+    describe("when get joinable rooms response from the docs is parsed") {
         val result = testFileReader.readTestFile("get_joinable_rooms_response-docs.json")
                 .parseAs<JoinableRoomsResponse>().successOrThrow()
 
-        it("then the result has the expected rooms") {
+        it("then the result has expected room") {
             val expectedRoom = NotJoinedRoomApiType(
-                    "ac43dfef",
-                    "alice",
-                    "Chatkit chat",
-                    null,
-                    mapOf("highlight_color" to "blue"),
-                    "2017-03-23T11:36:42Z",
-                    "2017-03-23T11:36:42Z",
-                    null)
+                    id = "ac43dfef",
+                    createdById = "alice",
+                    name = "Chatkit chat",
+                    pushNotificationTitleOverride = null,
+                    customData = mapOf("highlight_color" to "blue"),
+                    createdAt = "2017-03-23T11:36:42Z",
+                    updatedAt = "2017-03-23T11:36:42Z",
+                    deletedAt = null)
 
             assertThat(result.rooms).containsExactly(expectedRoom)
         }
